pyproject.toml
[project]
name = "vibe-coding-risk"
version = "0.1.0"
description = "Security risk analysis pipeline for vibe-coding chat datasets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "orjson>=3.10.0",
  "jsonschema>=4.23.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "tqdm>=4.67.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8.0",
  "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_configs = true

[tool.uv]
package = false
